Update README.md adding comment comparing ZHA Device Handlers to Zigbee-Shepherd Converters
Update README.md adding comment comparing ZHA Device Handlers to that of zigbee-shepherd-converters as used by the Zigbee2mqtt project in addition to the existing comparison to that of Hub-connected Device Handlers for the SmartThings Classics platform.

Hope that is a good and fair comparison? 

By the way, the Zigbee2mqtt project has written a great guide on how to add support for new devices by creating Zigbee-Shepherd Converters for the Zigbee2mqtt project. See http://www.zigbee2mqtt.io/how_tos/how_to_support_new_devices.html Maybe it could be a good idea to copy ideas from that guide to make a similar guide for how to create ZHA Device Handlers for Home Assistant?