xiaomi: send event when "model" attribute is reported
Some devices, like xiaomi sensors, send the model
attribute when their reset button is pressed quickly.
Having this report propagated to HA as an event
allows having a sound or light blink when a sensor
button is pressed, which is useful to make sure the
sensor is working. This behavior is similar to the
one provided by the Xiaomi Hub (that talks).