Changed the way the args were being parsed to prevent duplicate triggers.
Due to the way that the args are matched in HA, the current
implementation meant that the left and right button were each firing
triggers for both themself, and the other. So left would fire
automations triggered on left and right, and vice versa.

I've changed the "press" and "hold" command parsing to start with a
16 bit int, the same as release, this prevents the zeros being
generated for both, which was spooking the way that HA matches these.