Hue White Dimmable - Expose PhilipsLevelControlCluster (#556)
* Hue White Dimmable - Expose PhilipsLevelControlCluster

In order for PhilipsOnOff power restore to work, you also need to be able to set Level Control cluster 0x0008, PowerOn Level attribute 0x4000 to value 255 (0xff).

I edited the quirk to add the Level Control cluster like above. When I read/get to attribute ox4000, it is set as 254 as default. This causes the bulb to restore to full brightness after power restore. If set as 255, it restores to previous brightness. 

Findings based on this https://github.com/dresden-elektronik/deconz-rest-plugin/issues/1055#issuecomment-559993258

* Update zhaquirks/philips/zlldimmablelight.py

Co-authored-by: Alexei Chetroi <lexoid@gmail.com>

* Change philips import to be one line

Co-authored-by: Alexei Chetroi <lexoid@gmail.com>