Fixed factor/divider for _TZE200_b6wax7g0 temperature settings
For moeshouse ZTRV-BY-100-WH-DE thermostat radiator valve, the setpoint temperature was wrong by a factor of 10.
To correct this, I introduced 3 new classes for this device.
The main change are these 2 lines:
MOES_TARGET_TEMP_ATTR: ("occupied_heating_setpoint", lambda value: value * 100)
DIRECT_MAPPING_ATTRS = { "occupied_heating_setpoint": (MOES_TARGET_TEMP_ATTR, lambda value: round(value / 100),   ),
In the original code, the factor/divider was 10 which I changed to 100.
As a side effect, the temperature can only be set in steps of 1 degree and not in steps of 0.5 degrees. The reason for this is not clear. But it seems that in the cluster - which is an int value - the temperature is stored in degrees and not in centidegrees.