Explicitly enable "OppleMode" now (set to 1 instead of 0)
On v41 firmware, this attribute doesn't exist anymore. The plug behaves like it's set to "1"
On v32 firmware, this attribute exists. Now, that we have a proper quirk that parses Xiaomi attributes, we want this set to "1" (enabled). This causes older firmware plugs to behave like newer plugs always do.