Add missing long-press action to one TS0043 variant
While I don't have access to this variant, the missing long press actions seem
like an accidental omission (the PR #673 introducing this variant makes no
hint about it, and it links to an AliExpress description of the switch which
mentions long presses explicitly). Adding them seems harmless enough: the
worst case scenario is that these devices really don't generate long press
events, but then no real harm is done. Perhaps some users will be disappointed
to be shown this event as an option but not being able to actually use it. In
that case, my suggestion would be to wait for an issue report on this; then
these events can be removed again, but this time with a comment explicitly
stating this omission is intentional. In contrast, in the best case some users
will be happy to finally be able to use long press actions on their switches.