Add soil moisture class
Merged _soil.py into the general sensor file, and moved humidity reporting into the soil moisture class.

Have retained the ability to add a RH_Multiplier to the value to the humidity just in case its required by future devices. The _TZE200_myd45weu requires a 100 multiplier, which is the default.