Remove also updating attribute cache
With https://github.com/home-assistant/core/pull/89481, ZHA will do this when it gets a "zone status change command". We want to avoid doing this twice to cause one less state update in HA.

Before this PR, we also didn't update the ZONE_STATUS attribute cache (as the ZONE_STATE attribute was mistakenly updated).
So the PR mentioned above isn't necessary for this to work.