Add test for Aqara E1 roller curtain battery reporting
This device is a bit special, as it reports battery percentage via a Xiaomi Attribute and likely also voltage via Xiaomi attribute.
If it reports Xiaomi voltage too, we want to make sure that we do not use that voltage for calculating the battery percentage value.
We want it to use the battery percentage value sent from the device.