Revert to using custom logger (instead of zigpy logger mixin) for now
While the zigpy "mixin logger" includes useful information such as the NwkId, it also makes it seem like the message comes from zigpy (instead of zha-quirks).
Some of the quirks messages even include the IEEE of the device. Otherwise, enabling zigpy debug messages too should make it clear which device sent the message that produced the quirks log.

This behavior doesn't need to be changed in this PR, and we should consider to streamline it across zha-quirks in the future.