Move comment regarding `AnalogInputCluster` for current power consumption
Some devices have the `AnalogInput` cluster for current power consumption on endpoint 2 or even endpoint 1. Newer devices seem to always use endpoint 21 though.

The comment was also moved into the explanation of the docstring, as it rather belongs there.
(For this cluster, it doesn't matter on which endpoint it's used on -- so it won't "break". It's just an explanation of where it should likely be used.
Some devices might have multiple `AnalogInput` clusters, but only one should have this cluster in the replacement.)