Clean up Xiaomi quirks by removing `Bus` usage (#2528)
This PR refactors the Xiaomi quirks to remove Xiaomi specific usage of `Bus`
Other smaller changes:
- uses zigpy `AttributeDefs` in more places for getting attribute IDs (/removes "magic numbers")
- removes redundant `cluster_id` for custom clusters extending ZCL clusters
- improves one test by no longer calling the Bus directly, but instead creating a Xiaomi attribute report (new method `create_aqara_attr_report` will very likely be used in future PRs)
- adds test for Aqara P1 sensor
- adds test for old Aqara motion sensor
- adds test for Aqara weather sensor
- general cleanup 


* Create `create_aqara_attr_report` test method, remove `bus.listener_event` for MAEU01 plug in tests

* Remove `battery_bus`

* Remove `temperature_bus`

* Remove `humidity_bus`

* Remove `pressure_bus`

* Remove `voltage_bus`, `consumption_bus`, `power_bus`

* Remove `power_bus_percentage`

* Remove empty constructors

* Remove redundant `cluster_id`

* Small docs fixes

* Fix missing `self.ATTR_ID` for measurement clusters

* Use `AttributeDefs` when possible

* Remove unused strings

* Remove `Bus()` in constructor

* Get `zone_type` attr id from `AttributeDefs`

* Don't "calculate" value twice for total power consumption

* Fix missing `self.ATTR_ID` for Aqara P1 sensor

* Fix no super illuminance_reported() for Aqara P1 sensor

* Fix illuminance_reported() for Aqara P1 sensor on motion attr

* Use `self.debug` for Aqara P1 sensor, remove logger

* Use `self.AttributeDefs` from IlluminanceMeasurement, fix and/or condition

* Use `self.warning` for Xiaomi `__init__.py` instead of logger

* Add Aqara P1 motion sensor test

* Use `self.AttributeDefs`for Xiaomi `__init__.py`, as it's shorter

* Fix AnalogInputCluster forwarding current power to other endpoint

* Improve tests for Xiaomi plugs

* Check that quirk uses XiaomiPowerConfiguration before updating mV report

* Change non-working subclass check to AttributeError suppress

* Add test for Aqara weather sensors

* Fix typo

* Use AttributeDefs for P1 sensor

* Add test for old Aqara motion sensors

* Replace `attributes_by_name` with `AttributeDefs` in Xiaomi tests

* Add debug message when Xiaomi battery report is made, but cluster not used

* Add test which checks for debug message for devices reporting voltage, but not using XiaomiPowerConfiguration cluster

* Use `XiaomiPowerConfiguration` for roller curtain E1
and remove `PowerConfigurationRollerE1`

* Don't use voltage for (E1 roller curtain) calculating battery percentage

The device might send both. We only want to use voltage for the voltage attribute and the reported battery percentage for the percentage attribute. That's handled in `XiaomiPowerConfiguration`

Overriding `_update_battery_percentage` lets us ignore just the voltage for calculating percentage.

* Expand Aqara E1 motion sensor tests

* Use `update_attribute` instead of `_update_attribute` in tests

(doesn't really matter, but mostly used elsewhere, and it's just a method which calls `_update_attribute`)

* Add test for Aqara E1 roller curtain battery reporting

This device is a bit special, as it reports battery percentage via a Xiaomi Attribute and likely also voltage via Xiaomi attribute.
If it reports Xiaomi voltage too, we want to make sure that we do not use that voltage for calculating the battery percentage value.
We want it to use the battery percentage value sent from the device.

* Revert to using custom logger (instead of zigpy logger mixin) for now

While the zigpy "mixin logger" includes useful information such as the NwkId, it also makes it seem like the message comes from zigpy (instead of zha-quirks).
Some of the quirks messages even include the IEEE of the device. Otherwise, enabling zigpy debug messages too should make it clear which device sent the message that produced the quirks log.

This behavior doesn't need to be changed in this PR, and we should consider to streamline it across zha-quirks in the future.

* Use `AttributeDefs` for getting `ZONE_STATUS` attr id

This is used everywhere else already. `ZONE_STATUS` was/is special, because `zhaquirks.const` defines it too. (This is still used in other places, but should also be "removed" eventually probably.)

* Move comment regarding `AnalogInputCluster` for current power consumption

Some devices have the `AnalogInput` cluster for current power consumption on endpoint 2 or even endpoint 1. Newer devices seem to always use endpoint 21 though.

The comment was also moved into the explanation of the docstring, as it rather belongs there.
(For this cluster, it doesn't matter on which endpoint it's used on -- so it won't "break". It's just an explanation of where it should likely be used.
Some devices might have multiple `AnalogInput` clusters, but only one should have this cluster in the replacement.)