Rework implementation to update attribute again if doubled wrongly
We always double the battery pct first if the firmware is unknown (for devices which have the doubling clusters, very new IKEA devices do not at all).
But if the firmware is unknown, we cause a read to be sent (as the device is hopefully awake).
If that read now returns a new firmware, we "undouble" the battery pct again, so it's correct.

With the previous commit, it would only be correct if the device sends another update.
With this commit, we have a short flash of the doubled battery pct, then go to the correct value IF the device responded to our read.