Always read `sw_build_id` on `bind()`
This is for reading the attribute on a new join/repairing/reconfiguration, or refreshing the attribute in the two later cases.
The `sw_build_id` attribute is not updated after zigpy performed an OTA update. It will still be in cache, so our existing code won't update it in that case, only a reconfiguration will with this code.