Update ts000x.py
Change device type of Switch_4G_GPP to be an actual switch, as the device consists of potential-free relais. Also the type switch is substitutable with other types using HA helpers, but not (or not easily) vice versa.

It would probably make sense to change this for other similar switches as well, but as I don't own any others, I don't want to change stuff blindy, potentially introducing unnecessary errors/misbehaviour.