Add `_VALID_ATTRIBUTES` to `LocalDataCluster`
This list should be populated with the attribute ids of attributes that will not be populated at first, but only later.
All attribute reads on the `LocalDataCluster` for those specified attributes will return `None` with success status when no value is in the attribute cache.
This allows ZHA entity creation for those entities (except some configuration entities that explicitly check for `None` attributes).

If they are not in the valid attributes list, they will retain the previous behavior returning as "unsupported attribute".
This prevents ZHA entity creation for those attributes.