Create SONOFF S60ZBTPF quirk
The power and current values of the SONOFF S60ZBTPF and S60ZBTPG are not correctly reset to zero when the switch of the socket is turned off. Similarly, these devices sometimes even send power and current updates even after the switch was turned off.

This commit creates a custom quirk for the two devices to fix the broken behavior. I've only tested this with the SONOFF S60ZBTPF devices I have, but, judging by https://github.com/Koenkk/zigbee-herdsman-converters/blob/v25.44.0/src/devices/sonoff.ts#L2020, this should also apply to the S60ZBTPG.